Review
Renamed the nav_tui.py to navigator_status_tui and made it executable. It can now be run with:
>rosrun navigator_tools navigator_status_tui
I added a 2 hz rospy rate for the while loop which waits with a sleep() method.
I added a if condition that checks to see if the message sent over the '/wrench/selected' topic has '/wrench/' in the string and trims it if it does.