Review
I was confused about why the script wasn't executable for you as when I ran: 
> rosrun navigator_tools navigator_status_tui
the tui seemed to work as intended. I figure that this may be a permissions issue so I ran:
>chmod a+x navigator_status_tui
while in the uilts/navigator_tools/nodes/ directory to give executable permission to everyone.
Lemme know if the problem persists.